+++ ExpandWorld/vegetation/VegetationSpawning.cs
@@ -1,0 +1,122 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Emit;
+using HarmonyLib;
+using UnityEngine;
+
+namespace ExpandWorld;
+
+[HarmonyPatch(typeof(ZoneSystem), nameof(ZoneSystem.PlaceVegetation))]
+public class VegetationSpawning
+{
+  public static Dictionary<ZoneSystem.ZoneVegetation, Range<Vector3>> Scale = new();
+  public static Dictionary<ZoneSystem.ZoneVegetation, ZDO> ZDO = new();
+  private static ZoneSystem.ZoneVegetation Veg = new();
+  private static ZoneSystem.SpawnMode Mode = ZoneSystem.SpawnMode.Client;
+  private static List<GameObject> SpawnedObjects = new();
+  static void Prefix(ZoneSystem.SpawnMode mode, List<GameObject> spawnedObjects)
+  {
+    Mode = mode;
+    SpawnedObjects = spawnedObjects;
+  }
+  static ZoneSystem.ZoneVegetation SetVeg(ZoneSystem.ZoneVegetation veg)
+  {
+    Veg = veg;
+    return veg;
+  }
+  static void SetData(GameObject prefab, Vector3 position, Quaternion rotation, Vector3 scale, ZDO? data = null)
+  {
+    if (data == null)
+    {
+      if (!ZDO.TryGetValue(Veg, out data)) return;
+    }
+    if (data == null) return;
+    if (!prefab.TryGetComponent<ZNetView>(out var view)) return;
+    Data.InitZDO(position, rotation, scale, data, view);
+  }
+  static GameObject Instantiate(GameObject prefab, Vector3 position, Quaternion rotation)
+  {
+    SetData(prefab, position, rotation, Vector3.one);
+    var obj = UnityEngine.Object.Instantiate<GameObject>(prefab, position, rotation);
+    Data.CleanGhostInit(obj);
+    return obj;
+  }
+  static GameObject InstantiateWithData(GameObject prefab, Vector3 position, Quaternion rotation, Vector3 scale, ZDO? data = null)
+  {
+    SetData(prefab, position, rotation, scale, data);
+    var obj = UnityEngine.Object.Instantiate<GameObject>(prefab, position, rotation);
+    Data.CleanGhostInit(obj);
+    return obj;
+  }
+  static GameObject InstantiateBlueprint(GameObject prefab, Vector3 position, Quaternion rotation)
+  {
+    SetData(prefab, position, rotation, Vector3.one);
+    var obj = UnityEngine.Object.Instantiate<GameObject>(prefab, position, rotation);
+    Data.CleanGhostInit(obj);
+    if (BlueprintManager.TryGet(prefab.name, out var bp))
+      SpawnBlueprint(bp, position, rotation);
+    return obj;
+  }
+  static void SpawnBlueprint(Blueprint bp, Vector3 pos, Quaternion rot)
+  {
+    var zs = ZoneSystem.instance;
+    foreach (var obj in bp.Objects)
+    {
+      if (obj.Chance < 1f && UnityEngine.Random.value > obj.Chance) continue;
+      SpawnBPO(zs, pos, rot, Mode, SpawnedObjects, obj);
+    }
+  }
+  static void SpawnBPO(ZoneSystem zs, Vector3 pos, Quaternion rot, ZoneSystem.SpawnMode mode, List<GameObject> spawnedGhostObjects, BlueprintObject obj)
+  {
+    var objPos = pos + rot * obj.Pos;
+    var objRot = rot * obj.Rot;
+    var prefab = ZNetScene.instance.GetPrefab(obj.Prefab);
+    if (!prefab)
+    {
+      if (BlueprintManager.TryGet(obj.Prefab, out var bp))
+      {
+        SpawnBlueprint(bp, objPos, objRot);
+        return;
+      }
+      ExpandWorld.Log.LogWarning($"Blueprint prefab {obj.Prefab} not found!");
+      return;
+    }
+    if (mode == ZoneSystem.SpawnMode.Ghost)
+    {
+      ZNetView.StartGhostInit();
+    }
+    var go = InstantiateWithData(prefab, objPos, objRot, obj.Scale, obj.Data);
+    if (mode == ZoneSystem.SpawnMode.Ghost)
+    {
+      spawnedGhostObjects.Add(go);
+      ZNetView.FinishGhostInit();
+    }
+  }
+  static void SetScale(ZNetView view, Vector3 scale)
+  {
+    if (Scale.TryGetValue(Veg, out var randomScale))
+      scale = Helper.RandomValue(randomScale);
+    view.SetLocalScale(scale);
+  }
+  static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
+  {
+    var instantiator = typeof(UnityEngine.Object).GetMethods().First(m => m.Name == nameof(UnityEngine.Object.Instantiate) && m.IsGenericMethodDefinition && m.GetParameters().Skip(1).Select(p => p.ParameterType).SequenceEqual(new[] { typeof(Vector3), typeof(Quaternion) })).MakeGenericMethod(typeof(GameObject));
+    return new CodeMatcher(instructions)
+      .MatchForward(false, new CodeMatch(OpCodes.Ldfld, AccessTools.Field(typeof(ZoneSystem.ZoneVegetation), nameof(ZoneSystem.ZoneVegetation.m_enable))))
+      .Insert(new CodeInstruction(OpCodes.Call, Transpilers.EmitDelegate(SetVeg).operand))
+      .MatchForward(false, new CodeMatch(OpCodes.Call, instantiator))
+      .Set(OpCodes.Call, Transpilers.EmitDelegate(Instantiate).operand)
+      .MatchForward(false, new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(ZNetView), nameof(ZNetView.SetLocalScale))))
+      .Set(OpCodes.Call, Transpilers.EmitDelegate(SetScale).operand)
+      .MatchForward(false, new CodeMatch(OpCodes.Call, instantiator))
+      .Set(OpCodes.Call, Transpilers.EmitDelegate(InstantiateBlueprint).operand)
+      .InstructionEnumeration();
+  }
+}
+
+
+[HarmonyPatch(typeof(ZoneSystem), nameof(ZoneSystem.ValidateVegetation))]
+public class ValidateVegetation
+{
+  static bool Prefix() => false;
+}